requirements.txt
streamlit>=1.24.0
pandas>=1.5.0
numpy>=1.21.0
yfinance>=0.2.0
requests>=2.28.0
python-dotenv>=1.0.0
plotly>=5.13.0
scipy>=1.9.0
textblob>=0.17.1
vaderSentiment>=3.3.2